Enable Travis CI
Travis CI will automatically build the firmware and software projects
and report success or failure. This will apply to pull requests and
branches. The master branch should never be in a state where it doesn't
build. The goal is that master should always be stable and Travis CI
helps make this possible.